% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_batch_megadetector.R
\name{py_batch_megadetector}
\alias{py_batch_megadetector}
\title{Batch MegaDetector}
\usage{
py_batch_megadetector(
  image_path = "C:/MegaDetector/test_images/",
  out_file = "C:/MegaDetector/results/results.json",
  detector_model = "default",
  threshold = 0.1,
  ncores = 0,
  cp_freq = -1,
  recursive = TRUE,
  resume = FALSE
)
}
\arguments{
\item{image_path}{A file path to a folder containing camera trap images.}

\item{out_file}{A file path naming the output JSON file. Must end with
*.json. Note that checkpoint JSON files will be saved in this file path.}

\item{detector_model}{A file path to a trained detector model, must end with
*.pb. Microsoft's most current version-- md_v4.1.0.pb --is the internal
default (detector_model = NULL).}

\item{threshold}{The confidence threshold of the detector. Must be a number
between 0 and 1.0, boxes below this threshold will not be included in
output JSON file. Default value is 0.1.}

\item{ncores}{Number of CPU cores to use. Must be an integer. If > 1,
checkpointing will not be supported. Default value is 0.}

\item{cp_freq}{Frequency of saving a temporary file 'checkpoint'. Must be an
integer. Default value is -1, which disables checkpointing.}

\item{recursive}{A Boolean statement whether to recurse into a directory.
Default is TRUE.}

\item{resume}{A file path to a JSON checkpoint from which to resume the
function. Must be located in the same directory as the 'out_file'
parameter. Default is FALSE to ignore.}
}
\value{
A JSON file is saved in the out_file location.
}
\description{
Executes Microsoft's megadetector batch Python scipt within R using
'reticulate'. Only returns an output *.json file with detections for use with
Timelapse Image Analyser software.
}
