% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_simple_megadetector.R
\name{py_simple_megadetector}
\alias{py_simple_megadetector}
\title{Simple MegaDetector}
\usage{
py_simple_megadetector(
  image_path = "C:/MegaDetector/test_images/",
  out_path = "C:/MegaDetector/results/",
  detector_model = "default",
  view_when_done = FALSE
)
}
\arguments{
\item{image_path}{A file path to a folder containing camera trap images.}

\item{out_path}{A file path to a folder for storing the output images and
results.json file.}

\item{detector_model}{A file path to a trained detector model, must end with
*.pb. Microsoft's most current version-- md_v4.1.0.pb --is the internal
default (detector_model = NULL).}

\item{view_when_done}{A boolean statement, whether to open a window to view
the results in 'out_path' parameter when completed. Default is FALSE.}
}
\value{
Images with bounding box drawn and a 'results.json' file are saved in
the folder defined by the 'out_path' parameter.
}
\description{
Executes Microsoft's simple megadetector Python script within R using
'reticulate'.Returns an output *.json file with detections for use with
Timelapse Image Analyser software as well as new images containing the
detection bounding boxes.
}
